chore(client): make npm package browser friendly
The default nodejs package couldn’t be used in commander, so this makes it browser friendly unless you pull the CLI stuff.
