+++ client/nodejs/.gitignore
@@ -1,1 +1,2 @@
 lib
+./key*
