+++ client/nodejs/package.json
@@ -48,23 +48,23 @@
     }
   },
   "dependencies": {
-    "@commander-js/extra-typings": "^13.1.0",
     "@polkadot/api": "^15.8.1",
     "@polkadot/keyring": "^13.4.3",
     "@polkadot/util-crypto": "^13.4.3",
     "bignumber.js": "^9.1.2",
-    "commander": "^13.1.0",
     "console-table-printer": "^2.12.1",
-    "dotenv": "^16.4.7",
-    "typed-emitter": "^2.1.0"
+    "nanoevents": "^9.1.0"
+  },
+  "optionalDependencies": {
+    "@commander-js/extra-typings": "^13.1.0",
+    "commander": "^13.1.0",
+    "dotenv": "^16.4.7"
   },
   "devDependencies": {
     "@argonprotocol/testing": "1.1.0-rc.4",
-    "@microsoft/api-extractor": "^7.52.3",
     "@polkadot/typegen": "^15.8.1",
     "@substrate/ss58-registry": "^1.49.0",
     "@types/node": "^18.19.6",
-    "@vitest/coverage-v8": "3.1.1",
     "shx": "^0.4.0",
     "tsup": "^8.4.0",
     "tsx": "^4.19.2",
