+++ lerna.json
@@ -2,9 +2,5 @@
   "version": "1.1.0-rc.4",
   "npmClient": "yarn",
   "useWorkspaces": true,
-  "workspaces": [
-    "client/nodejs",
-    "localchain",
-    "localchain/npm/*"
-  ]
+  "workspaces": ["client/nodejs", "localchain", "localchain/npm/*"]
 }
