+++ runtime/common/src/lib.rs
@@ -89,7 +89,7 @@ macro_rules! inject_runtime_vars {
 			// `spec_name`,   `spec_version`, and `authoring_version` are the same between Wasm and
 			// native. This value is set to 100 to notify Polkadot-JS App (https://polkadot.js.org/apps) to use
 			//   the compatible custom types.
-			spec_version: 118,
+			spec_version: 119,
 			impl_version: 6,
 			apis: RUNTIME_API_VERSIONS,
 			transaction_version: 3,
